Refactored decompress and archives plugins into a component
